REFACTOR : DDD Layered Architecture 구성
