--- src/main/java/com/texturized/muus/common/error/ErrorCode.java
@@ -1,30 +1,0 @@
-package com.texturized.muus.common.error;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonFormat.Shape;
-import lombok.Getter;
-
-@Getter
-@JsonFormat(shape = Shape.OBJECT)
-public enum ErrorCode {
-
-    // Common
-    INVALID_INPUT_VALUE(400, "C001", " Invalid Input Value"),
-    METHOD_NOT_ALLOWED(405, "C002", " Unsupported HTTP Method"),
-    ENTITY_NOT_FOUND(400, "C003", " Entity Not Found"),
-    INTERNAL_SERVER_ERROR(500, "C004", " Server Error"),
-    INVALID_TYPE_VALUE(400, "C005", " Invalid Type Value"),
-    HANDLE_ACCESS_DENIED(403, "C006", " Access is Denied")
-
-    ;
-
-    private final int status;
-    private final String code;
-    private final String message;
-
-    ErrorCode(int status, String code, String message) {
-        this.status = status;
-        this.code = code;
-        this.message = message;
-    }
-}
