--- src/main/java/com/texturized/muus/common/response/Existence.java
@@ -1,8 +1,0 @@
-package com.texturized.muus.common.response;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class Existence {
-    private final boolean existence;
-}
