--- src/main/java/com/texturized/muus/common/resttemplate/SwaggerConfig.java
@@ -1,25 +1,0 @@
-package com.texturized.muus.common.resttemplate;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-            .components(new Components())
-            .info(apiInfo());
-    }
-
-    private Info apiInfo() {
-        return new Info()
-            .title("µ'us API")
-            .description("Springdoc을 사용한 Swagger UI 테스트")
-            .version("1.0.0");
-    }
-}
