--- src/main/java/com/texturized/muus/domain/user/dto/UserResponse.java
@@ -1,10 +1,0 @@
-package com.texturized.muus.domain.user.dto;
-
-import com.texturized.muus.domain.user.domain.User;
-
-public class UserResponse {
-
-    public UserResponse(final User user) {
-
-    }
-}
