--- src/main/java/com/texturized/muus/domain/user/exception/UserNotFoundException.java
@@ -1,10 +1,0 @@
-package com.texturized.muus.domain.user.exception;
-
-import javax.persistence.EntityNotFoundException;
-
-public class UserNotFoundException extends EntityNotFoundException {
-    public UserNotFoundException(Long target) {
-        super(target + " is not found");
-    }
-
-}
