+++ src/main/java/kr/texturized/muus/common/error/exception/BusinessException.java
@@ -1,0 +1,21 @@
+package kr.texturized.muus.common.error.exception;
+
+import kr.texturized.muus.common.error.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public BusinessException(String message, ErrorCode errorCode) {
+        super(message);
+        this.errorCode = errorCode;
+    }
+
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+
+}
