+++ src/main/java/kr/texturized/muus/common/response/Existence.java
@@ -1,0 +1,8 @@
+package kr.texturized.muus.common.response;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class Existence {
+    private final boolean existence;
+}
