+++ src/main/java/kr/texturized/muus/domain/exception/UserNotFoundException.java
@@ -1,0 +1,10 @@
+package kr.texturized.muus.domain.exception;
+
+import javax.persistence.EntityNotFoundException;
+
+public class UserNotFoundException extends EntityNotFoundException {
+    public UserNotFoundException(Long target) {
+        super(target + " is not found");
+    }
+
+}
