+++ src/main/java/kr/texturized/muus/presentation/api/response/UserResponse.java
@@ -1,0 +1,10 @@
+package kr.texturized.muus.presentation.api.response;
+
+import kr.texturized.muus.domain.entity.User;
+
+public class UserResponse {
+
+    public UserResponse(final User user) {
+
+    }
+}
