--- src/main/resources/com/texturized/muus/domain/user/dao/UserViewMapper.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.texturized.muus.domain.user.dao.UserViewMapper">
-
-  <sql id="select">
-    select user.id
-  </sql>
-
-  <sql id="from">
-    from
-        user
-  </sql>
-
-
-  <select id="getList" parameterType="map" resultType="com.texturized.muus.domain.user.domain.User">
-    <include refid="select"/>
-    <include refid="from"/>
-  </select>
-
-  <select id="getCount" parameterType="map" resultType="int">
-    select count(*)
-    <include refid="from"/>
-  </select>
-
-</mapper>
