+++ src/main/resources/import.sql
@@ -1,0 +1,3 @@
+
+INSERT INTO user(id) VALUES(1);
+INSERT INTO user(id) VALUES(2);
