+++ src/main/resources/kr/texturized/muus/infrastructure/mapper/UserViewMapper.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="kr.texturized.muus.infrastructure.mapper.UserViewMapper">
+
+  <select id="getList" parameterType="map" resultType="kr.texturized.muus.domain.entity.User">
+    select user.id
+    from user
+  </select>
+
+  <select id="getCount" parameterType="map" resultType="int">
+    select count(*)
+    from user
+  </select>
+
+</mapper>
