--- src/test/java/com/texturized/muus/MainTests.java
@@ -1,13 +1,0 @@
-package com.texturized.muus;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MainTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
