--- src/test/java/com/texturized/muus/test/config/TestProfile.java
@@ -1,5 +1,0 @@
-package com.texturized.muus.test.config;
-
-public interface TestProfile {
-    String TEST = "test";
-}
