+++ src/test/java/kr/texturized/muus/MainTests.java
@@ -1,0 +1,13 @@
+package kr.texturized.muus;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MainTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
