+++ src/test/java/kr/texturized/muus/test/config/TestProfile.java
@@ -1,0 +1,5 @@
+package kr.texturized.muus.test.config;
+
+public interface TestProfile {
+    String TEST = "test";
+}
